v1

+++ common/knnsearch.h
@@ -5,15 +5,15 @@
 class KNNSearch
 {
 public:
-    KNNSearch(float *data,int rows,int cols=3);
+    KNNSearch(float *data,int64_t rows,int64_t cols=3);
     ~KNNSearch(){if(flannIndex){ delete flannIndex; flannIndex=nullptr;}}
     std::vector<int> srcIndex;
 //    std::vector<int> dstIndex;
     std::vector<float> distances;
-    void search(float* dst,int dstRows,int dstCols);
+    void search(float* dst,int64_t dstRows,int64_t dstCols);
 private:
-    int rows;
-    int cols;
+    int64_t rows;
+    int64_t cols;
     std::vector<float> dataset;
     void reset();
     flann::Index<flann::MinkowskiDistance<float>>* flannIndex=nullptr;
